refactor: Hide database implementation behind an interface.
Here, an interface makes sense, since it's an I/O class. Hiding its
implementation makes it so changes to that implementation don't need a
rebuild of all its (50+) client code files.
