+++ CMakeLists.txt
@@ -64,6 +64,10 @@ project(qtox LANGUAGES CXX)
 # C++ standard.
 set(CMAKE_CXX_STANDARD 20)
 
+if(CMAKE_CXX_COMPILER_ID MATCHES "Clang")
+  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fcolor-diagnostics")
+endif()
+
 if(ASAN
    OR TSAN
    OR UBSAN)
@@ -374,6 +378,8 @@ set(${PROJECT_NAME}_SOURCES
     src/net/updatecheck.h
     src/persistence/db/rawdatabase.cpp
     src/persistence/db/rawdatabase.h
+    src/persistence/db/rawdatabaseimpl.cpp
+    src/persistence/db/rawdatabaseimpl.h
     src/persistence/db/upgrades/dbupgrader.cpp
     src/persistence/db/upgrades/dbupgrader.h
     src/persistence/db/upgrades/dbto11.h
