+++ src/BUILD.bazel
@@ -51,7 +51,7 @@ qt_moc(
         "net/toxuri.h",
         "net/updatecheck.h",
         "nexus.h",
-        "persistence/db/rawdatabase.h",
+        "persistence/db/rawdatabaseimpl.h",
         "persistence/history.h",
         "persistence/offlinemsgengine.h",
         "persistence/profile.h",
