+++ test/persistence/dbschema_test.cpp
@@ -154,19 +154,19 @@ void TestDbSchema::cleanup()
 void TestDbSchema::testCreation()
 {
     QVector<RawDatabase::Query> queries;
-    auto db = std::shared_ptr<RawDatabase>{new RawDatabase{testDatabaseFile->fileName(), {}, {}}};
+    auto db = RawDatabase::open(testDatabaseFile->fileName(), {}, {});
     QVERIFY(DbUpgrader::createCurrentSchema(*db));
     DbUtility::verifyDb(db, DbUtility::schema11);
 }
 
 void TestDbSchema::testIsNewDb()
 {
-    auto db = std::shared_ptr<RawDatabase>{new RawDatabase{testDatabaseFile->fileName(), {}, {}}};
+    auto db = RawDatabase::open(testDatabaseFile->fileName(), {}, {});
     bool success = false;
     bool newDb = DbUpgrader::isNewDb(db, success);
     QVERIFY(success);
     QVERIFY(newDb == true);
-    db = std::shared_ptr<RawDatabase>{new RawDatabase{testDatabaseFile->fileName(), {}, {}}};
+    db = RawDatabase::open(testDatabaseFile->fileName(), {}, {});
     createSchemaAtVersion(db, DbUtility::schema0);
     newDb = DbUpgrader::isNewDb(db, success);
     QVERIFY(success);
@@ -175,7 +175,7 @@ void TestDbSchema::testIsNewDb()
 
 void TestDbSchema::testNewerDb()
 {
-    auto db = std::shared_ptr<RawDatabase>{new RawDatabase{testDatabaseFile->fileName(), {}, {}}};
+    auto db = RawDatabase::open(testDatabaseFile->fileName(), {}, {});
     createSchemaAtVersion(db, DbUtility::schema0);
     int futureSchemaVersion = 1000000;
     db->execNow(
@@ -188,7 +188,7 @@ void TestDbSchema::testNewerDb()
 
 void TestDbSchema::test0to1()
 {
-    auto db = std::shared_ptr<RawDatabase>{new RawDatabase{testDatabaseFile->fileName(), {}, {}}};
+    auto db = RawDatabase::open(testDatabaseFile->fileName(), {}, {});
     createSchemaAtVersion(db, DbUtility::schema0);
     QVERIFY(DbUpgrader::dbSchema0to1(*db));
     DbUtility::verifyDb(db, DbUtility::schema1);
@@ -209,7 +209,7 @@ void TestDbSchema::test1to2()
     https://github.com/qTox/qTox/issues/5776
     */
 
-    auto db = std::shared_ptr<RawDatabase>{new RawDatabase{testDatabaseFile->fileName(), {}, {}}};
+    auto db = RawDatabase::open(testDatabaseFile->fileName(), {}, {});
     createSchemaAtVersion(db, DbUtility::schema1);
 
     const QString myPk = "AC18841E56CCDEE16E93E10E6AB2765BE54277D67F1372921B5B418A6B330D3D";
@@ -294,7 +294,7 @@ void TestDbSchema::test1to2()
 
 void TestDbSchema::test2to3()
 {
-    auto db = std::shared_ptr<RawDatabase>{new RawDatabase{testDatabaseFile->fileName(), {}, {}}};
+    auto db = RawDatabase::open(testDatabaseFile->fileName(), {}, {});
     createSchemaAtVersion(db, DbUtility::schema2);
 
     // since we don't enforce foreign key constraints in the db, we can stick in IDs to other tables
@@ -359,7 +359,7 @@ void TestDbSchema::test2to3()
 
 void TestDbSchema::test3to4()
 {
-    auto db = std::shared_ptr<RawDatabase>{new RawDatabase{testDatabaseFile->fileName(), {}, {}}};
+    auto db = RawDatabase::open(testDatabaseFile->fileName(), {}, {});
     createSchemaAtVersion(db, DbUtility::schema3);
     QVERIFY(DbUpgrader::dbSchema3to4(*db));
     DbUtility::verifyDb(db, DbUtility::schema4);
@@ -367,7 +367,7 @@ void TestDbSchema::test3to4()
 
 void TestDbSchema::test4to5()
 {
-    auto db = std::shared_ptr<RawDatabase>{new RawDatabase{testDatabaseFile->fileName(), {}, {}}};
+    auto db = RawDatabase::open(testDatabaseFile->fileName(), {}, {});
     createSchemaAtVersion(db, DbUtility::schema4);
     QVERIFY(DbUpgrader::dbSchema4to5(*db));
     DbUtility::verifyDb(db, DbUtility::schema5);
@@ -375,7 +375,7 @@ void TestDbSchema::test4to5()
 
 void TestDbSchema::test5to6()
 {
-    auto db = std::shared_ptr<RawDatabase>{new RawDatabase{testDatabaseFile->fileName(), {}, {}}};
+    auto db = RawDatabase::open(testDatabaseFile->fileName(), {}, {});
     createSchemaAtVersion(db, DbUtility::schema5);
     QVERIFY(DbUpgrader::dbSchema5to6(*db));
     DbUtility::verifyDb(db, DbUtility::schema6);
@@ -383,7 +383,7 @@ void TestDbSchema::test5to6()
 
 void TestDbSchema::test6to7()
 {
-    auto db = std::shared_ptr<RawDatabase>{new RawDatabase{testDatabaseFile->fileName(), {}, {}}};
+    auto db = RawDatabase::open(testDatabaseFile->fileName(), {}, {});
     // foreign_keys are enabled by History constructor and required for this upgrade to work on older sqlite versions
     db->execNow("PRAGMA foreign_keys = ON;");
     createSchemaAtVersion(db, DbUtility::schema6);
@@ -393,7 +393,7 @@ void TestDbSchema::test6to7()
 
 void TestDbSchema::test9to10()
 {
-    auto db = std::shared_ptr<RawDatabase>{new RawDatabase{testDatabaseFile->fileName(), {}, {}}};
+    auto db = RawDatabase::open(testDatabaseFile->fileName(), {}, {});
     createSchemaAtVersion(db, DbUtility::schema9);
 
     QVERIFY(insertFileId(*db, 1, true));
